cybrane_app/requirements.txt
fastapi
pydantic
qiskit-aer
qiskit-optimization
qiskit-algorithms
pymongo
python-dotenv
numpy
